+++ src/classes/features/NPCRemoval.ts
@@ -49,7 +49,10 @@ export class NPCRemoval extends RandomizerModFeature {
   ):
     | [entityType: EntityType, variant: int, subType: int, initSeed: Seed]
     | undefined {
-    return isOtherUnlockKindUnlocked(OtherUnlockKind.BLUE_FIREPLACES, true)
+    return isOtherUnlockKindUnlocked(
+      OtherUnlockKind.BLUE_AND_PURPLE_FIREPLACES,
+      true,
+    )
       ? undefined
       : [EntityType.FIREPLACE, FireplaceVariant.NORMAL, 0, initSeed];
   }
@@ -76,7 +79,10 @@ export class NPCRemoval extends RandomizerModFeature {
       return undefined;
     }
 
-    return isOtherUnlockKindUnlocked(OtherUnlockKind.PURPLE_FIREPLACES, true)
+    return isOtherUnlockKindUnlocked(
+      OtherUnlockKind.BLUE_AND_PURPLE_FIREPLACES,
+      true,
+    )
       ? undefined
       : [EntityType.FIREPLACE, FireplaceVariant.NORMAL, 0, initSeed];
   }